Adiciona maquina de estados funcional

+++ include/sensor_ultrassonico.h
@@ -17,7 +17,8 @@
 #define SOUND_SPEED 0.034
 #define CM_TO_INCH 0.393701
 
-int leituraUltrassonico();
+
+float leituraUltrassonico();
 
 void setupUltrassonico();
